Extract shared fuctionality from test server classes
