+++ src/test/java/com/yelp/nrtsearch/test_utils/TestFileUtils.java
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2025 Yelp Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.yelp.nrtsearch.test_utils;
+
+import java.io.IOException;
+import java.nio.file.DirectoryStream;
+import java.nio.file.Files;
+import java.nio.file.Path;
+
+public class TestFileUtils {
+
+  public static void rmDir(Path dir) throws IOException {
+    if (Files.exists(dir)) {
+      if (Files.isRegularFile(dir)) {
+        Files.delete(dir);
+      } else {
+        try (DirectoryStream<Path> stream = Files.newDirectoryStream(dir)) {
+          for (Path path : stream) {
+            if (Files.isDirectory(path)) {
+              rmDir(path);
+            } else {
+              Files.delete(path);
+            }
+          }
+        }
+        Files.delete(dir);
+      }
+    }
+  }
+}
